+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Skip unnecessary notification updates for redaction if the original notification does not exist.
+- Allow overriding existing mappings in EventContentSerializerMappings by type String.
 
 ### Deprecated